SImplify data model
Fusion RootFolderMixin and FolderMixin
Move RefentialObjectMixin in core (where it is only used)
Move updated_at in AbstracBaseModel
Remove redundant fields (UUID, folder)
